refactor(team): split prepare coordinator cache ports
